refac: handling buffers
Instead of hooking into buffer-list-update, partial-recall will now
advise partial-recall--switch-to-buffer-function which is set to
switch-to-buffer. This way the norecord flag of switch-to-buffer that
is also used by consult-buffer and the likes can be respected and
previewing buffers will not schedule handling such a buffer.

Adds find-file-hook to also handle that.