fix: remove scheduling from API commands
Adds new helper `partial-recall--new-buffer-p` that checks both the
old condition of not being the last checked buffer as well as checking
whether it is not set to be neglected.

The API commands now call new function
`partial-reclal--switch-to-buffer-and-neglect` that will set
`partial-recall--neglect` to that buffer.

This avoids scheduling buffers that were just inserted.