feat(history): add event structure, history methods
Adds new structure `partial-recall-event` that both moments and
memories can use. The event is identified by a type (symbol) and a
timestamp. It may optionally contain references to a moment and/or a
memory.

This also adds methods to record an event (generically) as well as to
retrieve events from a structure (optionally filtered by type).