improve(lighter): indicate if memory near capacity with contrast
This renames `partial-recall-memory-{at=>near}-capacity-p`. The
function now accepts optional parameter `threshold` that defaults to
0 (where it works as before). This allows checking how close to
capacity a memory is.

This new functionality is used in the memory part of the lighter. If
the memory size has grown to half of the default size, it is indicated
with face `contrast` instead of `deemphasized`.